\name{mdqc-package}
\alias{mdqc-package}
\docType{package}
\title{MDQC: Mahalanobis Distance Quality Control}

\description{
  MDQC is a multivariate quality assessment method for
  microarrays based on quality control (QC) reports.
}
\details{
\tabular{ll}{
Package: \tab mdqc\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2007-12-12\cr
License: \tab GPL\cr
}
}
\author{
  Justin Harrington \email{harringt@stat.ubc.ca} and Gabriela
  V. Cohen Freue \email{gcohen@stat.ubc.ca}.
}
\references{
  Cohen Freue, G. V. and Hollander, Z. and Shen, E. and Zamar, R. H. and Balshaw,
  R. and Scherer, A. and McManus, B. and Keown, P. and McMaster, W. R. and Ng,
  R. T. (2007) \sQuote{MDQC: A New Quality Assessment Method for Microarrays
  Based on Quality Control Reports}. \emph{Bioinformatics} \bold{23}, 3162 -- 3169.
}
\keyword{ package }
\seealso{\code{\link{mdqc}}}

